Block-sparse varlen attention for question branches
On CUDA (bf16/fp16) attention runs as a few flash varlen calls over the
whole packed batch instead of a dense masked call per request: trunks
causally on [past; trunk] keys, and every branch as (A) its segment's
trunk keys, unmasked, plus (B) its own keys, causal, merged exactly by
log-sum-exp. Cost follows what each branch reads rather than branch
tokens x all tokens (BRIGHT-like requests: ~97% of the dense work was
masked out). The index layout is built once per forward (Batch) and the
dense path stays as the CPU/fp32 path and the tests' reference; the tree
mask is now built only when that path needs it. Label logits are
computed in fp32.

+++ dan/models/llama.py
@@ -162,4 +162,4 @@ class LlamaForReads(nn.Module):
 
     def label_logits(self, h, label_ids):
         """Logits of just ``label_ids`` for one hidden state."""
-        return self.lm_head.weight[label_ids] @ h
+        return self.lm_head.weight[label_ids].float() @ h.float()  # fp32: bf16 logits step by 0.06+
